+++ Changelog.md
@@ -9,6 +9,20 @@
 - usage line refactor, cosmetic mostly
 - include \n in --version, cosmetic
 
+Migration guide:
+1. replace `positional_os("FOO")` with `positional("FOO").os()` or use `positional` in derive API
+2. replace `argument_os("FOO")` with `argument("FOO").os()` or use `argument` in derive API
+// 3. replace `impl Parser<T>` with `impl Parse<T>`
+// 4. replace `OptionParser<T>` with `Parser<T>`
+
+First two items are mostly to avoid combinatorial explosion
+
+You shouldn't be using those names directly, but there are some structure renames:
+1. `Positional` -> `ParsePositional`
+2. `BuildArgument` -> ParseArgument`
+3. `Command` -> `ParseCommand`
+
+
 ## bpaf [0.5.7] - 2022-09-04
 - bugfix with zsh autocomplete #46
 - reimplement bpaf derive - should be faster to compile and easier to work with